Add pagination for the main page. Set 5 posts on one page by default.
Create a PostDetailView view that returns detailed information about the post by the id field.

Add a template for this page and url with a path posts/pk/ and name post-detail.

On the post detail page, display a list of post comments below this post.

Under the list of comments, add a form that allows you to create a new comment to the post.